# openapi.yaml
openapi: 3.0.1
info:
  title: Priscillas Sample API
  description: watsonx Orchestrate testing.
  x-ibm-application-name: Priscilla Bot 0.0.1
  x-ibm-application-id: Priscilla-Bot-0.0.1
  x-ibm-application-icon: <svg id="Layer_1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 54 52.18"><path d="M12 21.593c-5.63-5.539-11-10.297-11-14.402 0-3.791 3.068-5.191 5.281-5.191 1.312 0 4.151.501 5.719 4.457 1.59-3.968 4.464-4.447 5.726-4.447 2.54 0 5.274 1.621 5.274 5.181 0 4.069-5.136 8.625-11 14.402m5.726-20.583c-2.203 0-4.446 1.042-5.726 3.238-1.285-2.206-3.522-3.248-5.719-3.248-3.183 0-6.281 2.187-6.281 6.191 0 4.661 5.571 9.429 12 15.809 6.43-6.38 12-11.148 12-15.809 0-4.011-3.095-6.181-6.274-6.181"/></svg>
  x-ibm-skill-type: imported
  x-ibm-skill-headers:
    caller-id: caller-id
    caller-name: caller-name
  x-ibm-comm-timeouts:
    connectTimeout: 20000
    responseTimeout: 120000
  version: 0.0.1

security:
  - oAuth2: []
servers:
  - url: https://python-api-production-f415.up.railway.app/
    description: Hosted on Railway.app

tags:
  - name: hello
    description: Greet the user
  - name: todo-list
    description: Get database of todo items

paths:
  /hello:
    get:
      summary: Greet the user
      description: Greet the user, greeting depends on the time of day
      tags:
        - hello
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/hello'
      # TODO: Change json msg into chat message using
      # x-ibm-nl-output-template: {$.msg}
  /item-list:
    get:
      summary: Get list of items
      description: Get list of items in the database
      tags:
        - item-list
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/item-list'
  /incident/{id}:
    get:
      summary: Get incident
      description: Get incident in ServiceNow
      security:
        - oAuth2: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
            title: sys_id
            # x-ibm-show: false
          required: true
          description: unique job id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/incident'
  /incident-list:
    get:
      summary: Get incidents
      description: Get all incidents from ServiceNow
      # security:
      #   - oAuth2: []
      responses:
        '200':
          description: OK
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/incident-list'
  /incident/create:
    post:
      summary: Create incident
      description: Create new incident in ServiceNow
      requestBody:
        content:
          application/json:
            schema:
              properties:
                  short_description:
                    type: string
                    
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/incident'

components:
  securitySchemes:
    oAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://dev251468.service-now.com/oauth_token.do
          scopes: {}
  schemas:
    hello:
      type: object
      properties:
        msg:
          type: string
    Item:
      type: object
      properties:
        id: 
          type: integer
        name:
          type: string
        priority:
          type: integer
    item-list:
      type: array
      items:
        $ref: '#/components/schemas/Item'
    sysID:
    #sysID is a required output for composite skills, not shown to user in table
      x-ibm-show: true
      type: string
    incident:
      type: object
      properties:
        sys_id:
          # $ref: '#/components/schemas/sysID'
          type: string
        opened_at:
          type: string
        assignment_group:
          type: string
        assigned_to:
          type: string
        urgency:
          type: string
        opened_by:
            type: string
            x-ibm-important: true
        state:
          type: string
        description:
          type: string
        short_description:
          type: string
          x-ibm-important: true
        number:
          type: string
    incident-list:
      type: array
      items:
        $ref: '#/components/schemas/incident'